+++ modules/emulator/measurements/s1.json
@@ -1,11 +1,11 @@
 [
   {
       "2": 100,
-      "12": 50
+      "12": 25
   },
 
   {
-      "2": 100,
+      "2": 25,
       "12": 50
   }
 ]
