--- modules/emulator/monitor.py
@@ -1,16 +1,0 @@
-
-# Handles fetching latency metrics from the measurement client
-class Monitor:
-    def __init__(self):
-        pass
-
-    def fetch_latency_results(self):
-
-        while True:
-            # Read Json Results from all nodes
-
-            # Perform algorithm Step
-
-            # Update flow tables at each node
-
-            pass
